update environment and application rb

+++ config/application.rb
@@ -17,9 +17,11 @@ module BookLibrary
   class Application < Rails::Application
     config.before_configuration do
       env_file = File.join(Rails.root, 'config', 'local_env.yml')
-      YAML.load(File.open(env_file)).each do |key, value|
-        ENV[key.to_s] = value
-      end if File.exists?(env_file)
+      if File.exists?(env_file)
+        YAML.load(File.open(env_file)).each do |key, value|
+          ENV[key.to_s] = value
+        end
+      end
     end
 
     # Settings in config/environments/* take precedence over those specified here.
@@ -29,7 +31,7 @@ module BookLibrary
     # Set Time.zone default to the specified zone and make Active Record auto-convert to this zone.
     # Run "rake -D time" for a list of tasks for finding time zone names. Default is UTC.
     config.time_zone = 'Pacific Time (US & Canada)'
-    
+
     config.action_mailer.default_url_options = { :host => ENV["DEFAULT_URL"] }
   end
 end
